Update write button component
