+++ apps/client/src/components/index.ts
@@ -1,3 +1,4 @@
 export { default as SigninButton } from './signinButton';
 export { default as Header } from './header';
 export { default as AlarmModal } from './alarmModal';
+export { default as WriteButton } from './writeButton';
